network: add command to protect wpa_supplicant config
By default everytime this file is saved it will set read permissions for all
users which makes the passwords for the Wi-Fi networks completely exposed.

Allow the wheel user to run this command without password to change the read
permissions to root only. This should be run in conjunction with the
`wpa_cli save_config` command.